modified taskcontroller

+++ app/Http/Controllers/TasksController.php
@@ -17,11 +17,20 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
+       $data = [];
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('id')->paginate(1000);
 
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+            // $data += $this->counts($user);
+            return view('tasks.index', $data);
+        }else {
+            return view('welcome');
+        }
     }
 
     /**
@@ -46,19 +55,25 @@ class TasksController extends Controller
      */
     public function store(Request $request)
     {
-        
         $this->validate($request, [
-            'status' => 'required|max:10',
             'content' => 'required|max:191',
+            'status' => 'required|max:10',
         ]);
-        
-        $task = new Task;
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
 
-        return redirect('/');
-    }
+        $request->user()->tasks()->create([ 
+             'content' => $request->content, 
+             'status' => $request->status, 
+         ]); 
+          
+         // $task = new Task; 
+         // $task->status = $request->status;  
+         // $task->content = $request->content; 
+         // $task->save(); 
+          
+  
+         return redirect('/tasks'); 
+     } 
+
 
     /**
      * Display the specified resource.
@@ -68,12 +83,19 @@ class TasksController extends Controller
      */
     public function show($id)
     {
-        $task = Task::find($id);
-
-        return view('tasks.show', [
-            'task' => $task,
-        ]);
+        { 
+         $task = Task::find($id); 
+          
+         if (\Auth::user()->id === $task->user_id){ 
+             return view('tasks.show', [ 
+             'task' => $task, 
+         ]); 
+         } else { 
+         return redirect('/'); 
+         } 
+     }
     }
+
 
     /**
      * Show the form for editing the specified resource.
@@ -81,14 +103,20 @@ class TasksController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function edit($id)
-    {
-        $task = Task::find($id);
+    public function edit($id) 
+     { 
+         $task = Task::find($id); 
+          
+         if (\Auth::user()->id === $task->user_id){ 
+             return view('tasks.edit', [ 
+             'task' => $task, 
+         ]); 
+         return redirect('/'); 
+              
+         } 
+     } 
 
-        return view('tasks.edit', [
-            'task' => $task,
-        ]);
-    }
+
 
     /**
      * Update the specified resource in storage.
@@ -120,10 +148,18 @@ class TasksController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function destroy($id)
-    {
-        $task = Task::find($id);
-        $task->delete();
-
-        return redirect('/');
+    { 
+         $task = \App\Task::find($id); 
+  
+         if (\Auth::user()->id === $task->user_id) { 
+             $task->delete(); 
+         return redirect('/'); 
+         } else { 
+  
+         //return redirect()->back(); 
+         return view('welcome'); 
+         } 
+     } 
+    
     }
-}
+
